requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1 # avoid compatibility issues with passlib  
bs4